feat: restructure for better design
