+++ mesibot-app/src/pages/Home/views/GameView/ShemKod/Instructions.tsx
@@ -1,38 +1,6 @@
-import { Button, styled } from "@mui/material";
-import { Colors } from "../../../../consts/colors";
-
-const PageTitle = styled("h2")`
-  direction: rtl;
-  font-size: 2rem; // Default font size
-
-  @media (max-width: 600px) {
-    font-size: 1.5rem; // Smaller screens (mobile)
-  }
-
-  @media (min-width: 1200px) {
-    font-size: 3rem; // Larger screens
-  }
-`;
-
-const Content = styled("p")`
-  text-align: center;
-  margin: 0 2vw 2vw 2vw;
-  direction: rtl;
-
-  font-size: 1.2rem; // Default font size
-
-  @media (max-width: 600px) {
-    font-size: 1rem; // Mobile-friendly text
-  }
-
-  @media (min-width: 1200px) {
-    font-size: 1.8rem; // Bigger text on large screens
-  }
-`;
-
-const Bold = styled("span")`
-  font-weight: bold;
-`;
+import { Button } from "@mui/material";
+import { Colors } from "../../../../../consts/colors";
+import { Bold, Content, PageTitle } from "../GameView.style";
 
 interface InstructionsProps {
   setStarted: (started: boolean) => void;
