+++ mesibot-app/src/pages/Home/views/GameView/ShemKod/index.ts
